added Skills/Interests Pages.
