+++ frontend/package.json
@@ -11,8 +11,10 @@
   },
   "dependencies": {
     "axios": "^1.6.7",
+    "multiselect-react-dropdown": "^2.0.25",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.0.1",
     "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
