+++ frontend/src/home/Home.jsx
@@ -4,7 +4,7 @@ const Home = () => {
 
             <div>
 
-                <h1>Home</h1>
+                <h1>Home Page</h1>
 
             </div>
