--- frontend/src/myprofile/Header.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-
-
-
-const Header = ({ image }) => {
-
-
-    return (
-        <div className="hear">
-                <img src={image} id='Pic' className="background-picture" />
-        </div>
-    
-
-    );
-};
-
-export default Header;
-
-
