+++ frontend/src/setting/settings.css
@@ -1,33 +1,34 @@
-.settings {
-    max-width: 400px; /* Adjust the width as needed */
-    margin: 0 auto;
-  }
-  
-  .settings h1 {
-    margin-bottom: 20px;
-  }
-  
-  .settings ul {
-    list-style-type: none;
-    padding: 0;
-  }
-  
-  .settings li {
-    position: relative;
-    padding-bottom: 10px;
-    margin-bottom: 10px;
-  }
-  
-  .settings li::after {
-    content: "";
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    height: 1px;
-    background-color: #ccc;
-  }
-  
-  .settings li:hover {
-    background-color: #f0f0f0;
-  }
+.settings ul {
+  list-style-type: none;
+  padding: 0;
+  align-items: center;
+}
+
+.settings li {
+  display: flex;
+  align-items: center;
+  padding: 10px;
+  cursor: pointer;
+ 
+}
+
+.settings li:hover {
+  background-color: #f0f0f0;
+}
+
+.settings li svg {
+  margin-right: 10px;
+}
+
+/* .settings{
+  align-items: center;
+} */
+
+
+
+
+
+
+/* https://ws24-skillswap.onrender.com/api/users/660e9b92453cad62e9a4f568 */
+
+/* where i can find all these icons for these skills */
